requirements.txt
certifi==2017.11.5
elasticsearch==6.0.0
google-api-core==0.1.1
google-auth==1.2.0
google-cloud-core==0.28.0
google-cloud-storage==1.6.0
googleapis-common-protos==1.5.3
